test refactoring T67625393
[ghstack-poisoned]
